+++ src/components/myProfile/myFollowingModal.tsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
-import ProfileCard from "./profileCard";
+import ProfileCard from "../profileCard";
 
 const Overlay = styled.div`
   position: fixed;
@@ -102,6 +102,16 @@ const Divider = styled.div`
     align-items: center; /* 가운데 정렬 */
 `
 
+const LoadingContainer = styled.div`
+    /* 로딩 메시지 스타일 추가 */
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    color: white;
+    font-size: 24px;
+`;
+
 interface FollowingModalProps {
     onClose: () => void;
 }
@@ -133,7 +143,7 @@ export default function MyFollowingModal ({ onClose } : FollowingModalProps) {
       
     return(
         <Overlay onClick={onClose}>
-            <Modal>
+            <Modal onClick={(e) => e.stopPropagation()}>
                 <Header>
                     <Title>팔로잉</Title>
                     <CloseButton onClick={onClose}>
@@ -144,7 +154,7 @@ export default function MyFollowingModal ({ onClose } : FollowingModalProps) {
                 <ModalContent>
                     <ProfileCard />
                     {loading ? (
-                        <div>Loading...</div>
+                        <LoadingContainer>Loading...</LoadingContainer>
                     ):(
                         followings && followings.map((following) => (
                             <ProfileCard
